requirements.txt
streamlit==1.36.0

pandas==2.1.4
numpy==1.24.4

python-dotenv==1.0.1

langchain==0.2.11
langchain-core==0.2.23
langchain-community==0.2.7
langchain-text-splitters==0.2.2
langchain-groq==0.1.4

faiss-cpu==1.7.4

ollama==0.1.9
sentence-transformers==2.7.0

tqdm==4.66.4
groq==0.8.0
